+++ ptools/cli/ptools_cli.py
@@ -5,26 +5,21 @@ For more specific help, provide the name of a positional argument, e.g.,
    ptools --help
 """
 
+import click
 
-import argparse
-
-from . import attract_cmd, reduce_cmd
-
-
-def parse_command_line(args=None):
-    """Main ptools command-line parsing."""
-    parser = argparse.ArgumentParser(description=__doc__)
-    parser.add_argument("--version", action="version")
-    subparsers = parser.add_subparsers(required=True)
-    attract_cmd.create_subparser(subparsers)
-    reduce_cmd.create_subparser(subparsers)
-    return parser.parse_args(args)
+from .attract_cmd import attract
+from .logger import configure_logger, logger
+from .reduce_cmd import reduce
 
 
-def main(args=None):
-    """Main ptools command."""
-    args = parse_command_line(args)
-    args.func(args)
+@click.group(help=__doc__)
+def main():
+    logger.enable("ptools")
+    configure_logger(debug=False)
+
+
+main.add_command(attract)
+main.add_command(reduce)
 
 
 if __name__ == "__main__":