+++ ptools/forcefield.py
@@ -1,11 +1,14 @@
 """Ptools forcefield implementation."""
 
+from typing import TYPE_CHECKING
+
 import numpy as np
 from scipy.spatial.distance import cdist
 
-from .attract import AttractRigidBody
-from .io.readers.attract import read_aminon
 from .pairlist import PairList
+
+if TYPE_CHECKING:
+    from .attract import AttractRigidBody
 
 # Name of the force fields implemented in pyattract.
 ATTRACT_FORCEFIELDS = ("scorpion", "attract1", "attract2")
@@ -64,22 +67,19 @@ ATTRACT_DEFAULT_FF_PARAMS = np.array(
 class AttractForceField1:
     """The AttractForceField1."""
 
-    receptor: AttractRigidBody
-    ligand: AttractRigidBody
+    receptor: "AttractRigidBody"
+    ligand: "AttractRigidBody"
     cutoff: float
-    paramfile: str
 
     def __init__(
         self,
-        receptor: AttractRigidBody,
-        ligand: AttractRigidBody,
+        receptor: "AttractRigidBody",
+        ligand: "AttractRigidBody",
         cutoff: float = 10,
-        paramfile: str = "",
     ):
         self.receptor = receptor
         self.ligand = ligand
         self.cutoff = cutoff
-        self.paramfile = paramfile
 
         self._vdw_energy = 0.0
         self._electrostatic_energy = 0.0
@@ -92,21 +92,14 @@ class AttractForceField1:
         self._initialize_parameters()
 
     def _initialize_parameters(self):
-        if self.paramfile != "":
-            params = read_aminon(self.paramfile)
-        else:
-            params = ATTRACT_DEFAULT_FF_PARAMS
+        params = ATTRACT_DEFAULT_FF_PARAMS
 
         rad, amp = list(zip(*params, strict=True))
         rad = np.array(rad)
         amp = np.array(amp)
 
-        self._repulsive_parameters = (
-            amp[:, None] * amp[:] * np.power(rad[:, None] + rad[:], 8)
-        )
-        self._attractive_parameters = (
-            amp[:, None] * amp[:] * np.power(rad[:, None] + rad[:], 6)
-        )
+        self._repulsive_parameters = amp[:, None] * amp[:] * np.power(rad[:, None] + rad[:], 8)
+        self._attractive_parameters = amp[:, None] * amp[:] * np.power(rad[:, None] + rad[:], 6)
 
         # Categorie pairs.
         C = np.array(np.meshgrid(self.receptor.typeids, self.ligand.typeids)).T
@@ -142,7 +135,6 @@ class AttractForceField1:
         """Private method for non-bonded energy calculation with small cutoffs."""
 
         def van_der_waals(dx, rr2):
-            # pylint: disable=E1126
             a = np.array([self._attractive_pairs[i, j] for i, j in zip(*keep, strict=True)])
             b = np.array([self._repulsive_pairs[i, j] for i, j in zip(*keep, strict=True)])
 
@@ -161,11 +153,7 @@ class AttractForceField1:
             return vlj.sum()
 
         def electrostatics(dx, rr2):
-            charge = (
-                self.receptor.charges[:, None]
-                * self.ligand.charges
-                * (332.053986 / 20.0)
-            )
+            charge = self.receptor.charges[:, None] * self.ligand.charges * (332.053986 / 20.0)
             charge = charge[keep]
 
             et = charge * rr2
@@ -210,11 +198,7 @@ class AttractForceField1:
             return vlj.sum()
 
         def electrostatics(dx, rr2):
-            charge = (
-                self.receptor.charges[:, None]
-                * self.ligand.charges
-                * (332.053986 / 20.0)
-            )
+            charge = self.receptor.charges[:, None] * self.ligand.charges * (332.053986 / 20.0)
             et = charge * rr2
             fdb = dx * (2.0 * et)[:, :, None]
             self.receptor.forces += fdb.sum(axis=1)
@@ -251,10 +235,7 @@ class AttractForceField1:
         contacts = pairlist.contacts()
         norm2 = pairlist.sqdistances()
 
-        alldx = [
-            self.receptor.coordinates[ir] - self.ligand.coordinates[il]
-            for ir, il in contacts
-        ]
+        alldx = [self.receptor.coordinates[ir] - self.ligand.coordinates[il] for ir, il in contacts]
 
         for i, (ir, il) in enumerate(contacts):
             category_rec = self.receptor.typeids[ir]