+++ HISTORY.rst
@@ -2,6 +2,24 @@
 History
 =======
 
+0.2.1 (2025-02-06)
+------------------
+
+* PDB readers reads alternate locations
+* logging using loguru
+* command-line interface uses click
+* various refactorings
+
+
+0.2.0 (2024-12-11)
+------------------
+
+* Version used by heligeom webserver v1.0.0
+* Major refactor of the codebase
+* Added new features
+* Improved performance
+
+
 0.1.0 (2016-10-17)
 ------------------